Controllers, more services and etc

+++ SmartGarage/Models/Service.cs
@@ -13,7 +13,7 @@ namespace SmartGarage.Models
 
         [Required]
         [Range(0, int.MaxValue)]
-        public decimal Price { get; set; }
+        public double Price { get; set; }
 
         public int UserId { get; set; }
